Add mock context manager to consolidate code paths
